+++ assets/src/commander/utils/bundlePlugin.js
