+++ assets/src/commander/utils/copyFiles.js
