+++ assets/src/commander/utils/deleteExistingZipFiles.js
