+++ assets/src/commander/utils/generateZipFile.js
