+++ assets/src/commander/utils/getCurrentVersionNumber.js
