+++ assets/src/commander/utils/getIgnoredFiles.js
