+++ assets/src/commander/utils/test/copyFiles.js
