+++ assets/src/commander/utils/test/deleteExistingZipFiles.js
