+++ assets/src/commander/utils/test/generateZipFile.js
