+++ assets/src/fonts/utils/test/buildFonts.js
