+++ assets/src/fonts/utils/test/fixtures/abezee.ttf
