+++ assets/src/fonts/utils/test/normalizeFont.js
