+++ assets/src/migration/utils/test/updateTemplates.js
