+++ assets/src/migration/utils/updateTemplates.js
@@ -22,7 +22,7 @@ import { readdirSync, readFileSync, writeFileSync } from 'fs';
 /**
  * Internal dependencies
  */
-import { migrate, DATA_VERSION } from '../../assets/src/migration/migrate.js';
+import { migrate, DATA_VERSION } from '../module.js';
 
 function updateTemplates(templatesDir) {
   const fileNamePattern = /^.*\.json$/;
