+++ package.json
@@ -228,10 +228,12 @@
     "test:e2e:firefox:interactive": "PUPPETEER_PRODUCT=firefox npm run test:e2e:interactive",
     "test:e2e:firefox:debug": "PUPPETEER_PRODUCT=firefox PUPPETEER_HEADLESS=false PUPPETEER_DEVTOOLS=true node --inspect-brk node_modules/.bin/jest --runInBand --config=tests/e2e/jest.config.js",
     "percy": "percy snapshot --config=percy.config.yml .test_artifacts/karma_snapshots/",
-    "workflow:version": "./bin/commander.js version",
-    "workflow:build-plugin": "./bin/commander.js build-plugin",
-    "workflow:fonts": "./bin/commander.js build-fonts",
-    "workflow:templates": "./bin/commander.js update-templates"
+    "workflow:version": "./assets/src/commander/index.js version",
+    "workflow:build-plugin": "./assets/src/commander/index.js build-plugin",
+    "workflow:fonts": "./assets/src/fonts/index.js assets/src/fonts/fonts.json",
+    "workflow:migrate": "npm-run-all --parallel workflow:templates:*",
+    "workflow:migrate:templates": "./assets/src/migration/cli.js assets/src/dashboard/templates/raw",
+    "workflow:migrate:ftue-story": "./assets/src/migration/cli.js includes/data/stories"
   },
   "husky": {
     "hooks": {
